Use promises to give messaging control to socket

+++ server/mazedraw.js
@@ -5,31 +5,54 @@ const config = require("./config");
 
 const log = require("./log").getLogger(module.filename);
 
-exports.draw = function (width, height, maze, cellSize, cb) {
+exports.draw = function (width, height, maze, cellSize) {
   const { walls, startingPoint, ends } = maze;
 
-  try {
-    let client = net.Socket();
-    const host = config.mazedraw.host,
-      port = config.mazedraw.port;
-    client.connect(port, host);
-    log.info(`Sending maze request to ${host}:${port}`);
-    client.write(
-      JSON.stringify({ width, height, walls, startingPoint, ends, cellSize })
-    );
+  return new Promise((resolve, reject) => {
+    try {
+      let client = net.Socket();
+      const host = config.mazedraw.host,
+        port = config.mazedraw.port;
+      client.connect(port, host);
+      log.info(`Sending maze request to ${host}:${port}`);
 
-    client.on("data", function (data) {
-      cb(data);
-      client.destroy();
-    });
-    client.on("error", function (ex) {
-      log.error("Could not connect with mazedraw service", ex);
-      cb(null);
-      client.destroy();
-    });
+      const timeoutMs = 1000 * 60; // 1 minute
 
-    client.end();
-  } catch (e) {
-    log.error("Encountered exception when connecting with mazedraw service", e);
-  }
+      let terminated = false;
+      // Set terminated if it's unset, and return the old value.
+      function setFlagAndGetOld() {
+        if (!terminated) {
+          terminated = true;
+          return false;
+        } else {
+          return true;
+        }
+      }
+
+      client.on("data", function (data) {
+        if (setFlagAndGetOld()) return;
+
+        resolve(data);
+        client.destroy();
+      });
+      client.on("error", function (ex) {
+        if (setFlagAndGetOld()) return;
+
+        reject(ex);
+        client.destroy();
+      });
+      setTimeout(function () {
+        if (setFlagAndGetOld()) return;
+
+        reject(new Error(`mazedraw service timed out at ${timeoutMs}ms`));
+      }, timeoutMs);
+
+      client.write(
+        JSON.stringify({ width, height, walls, startingPoint, ends, cellSize })
+      );
+      client.end();
+    } catch (ex) {
+      log.error("draw exception ", { ex });
+    }
+  });
 };
